Removed time range from worklet animation tests and updated expectations
Animation worklets do not yet support progress based timelines and still
rely on timeRange. However, for the sake of moving forward and because
animation worklets are still experimental it was deemed ok to break them
and update the test expectations until a time when they can be changed
to support progress based animations.

Bug: 1234315, 1216655, 1140602
Change-Id: I04e3ecb7c49cd2a93fe15f5c52f0acd2f18bc151