Don't let attachShadow() succeed subsequent times for declarative shadow roots
Currently, attachShadow() can be called any number of times for a
shadow host with a declarative shadow root, and each time the shadow
will be emptied out and returned.

The correct behavior is to only do this the first time attachShadow()
is called, and then throw errors an subsequent attempts. Implement
this behavior by clearing the declarative shadow root flag for
roots returned from Element::attachShadow.

Change-Id: I5d3d31fc8eaa09e5866191135743fe3f610accaa